Removed a file, disabled checkoutBtn empty cart
